DB API:  move API exports up one level, into server/db proper
