--- server/db/api/index.ts
@@ -1,11 +1,0 @@
-export * from './access';
-export * from './asset';
-export * from './capturedata';
-export * from './license';
-export * from './misc';
-export * from './model';
-export * from './systemobject';
-export * from './unit';
-export * from './user';
-export * from './vocabulary';
-export * from './workflow';
