+++ server/db/index.ts
@@ -3,3 +3,14 @@
  *
  * Organize and export db specific bindings here
  */
+export * from './api/access';
+export * from './api/asset';
+export * from './api/capturedata';
+export * from './api/license';
+export * from './api/misc';
+export * from './api/model';
+export * from './api/systemobject';
+export * from './api/unit';
+export * from './api/user';
+export * from './api/vocabulary';
+export * from './api/workflow';
